+++ tests/integration/test_reduce_REF_M_live_post_proc.py
@@ -2,17 +2,18 @@
 import itertools
 import os
 import shutil
-import string
+import unittest.mock as mock
 
 # third party imports
 import pytest
 
 # mr_reduction imports
-from mr_reduction.simple_utils import add_to_sys_path
+from mr_livereduce.reduce_REF_M_live_post_proc import main
+from selenium.webdriver.common.by import By
 
 
 @pytest.mark.datarepo()
-def test_template(mock_filesystem, data_server, autoreduction_script):
+def test_main(mock_filesystem, data_server, browser, autoreduction_script):
     r"""Substitute values in the template and then run a reduction using functions defined within the template
 
     Ideally, one would like to open a subprocess and invoke the reduction script, but it's not possible because
@@ -23,7 +24,7 @@ def test_template(mock_filesystem, data_server, autoreduction_script):
     # Gather all necessary auxiliary files for reduction of run 42537, which is the third run in the sequence
     # of runs that begins with run 42535.
     #
-    # direct beam for data run 42537
+    # direct beam for data run 41447
     mock_filesystem.DirectBeamFinder.return_value.search.return_value = 42534
     # autoreduced files from previous runs, to be stitched to profile from 42537
     for run, suffix in itertools.product(
@@ -35,40 +36,57 @@ def test_template(mock_filesystem, data_server, autoreduction_script):
 
     # Create a temporary autoreduction script reduce_REF_M.py and pass its parent directory to PYTHONPATH.
     # The default options in this script are good for reducing the two peaks of run 41447.
-    reduction_script = autoreduction_script()
+    autoreduction_script(outdir=mock_filesystem.tempdir)
 
-    # We don't invoke the reduction script as a shell command because we need mock_filesystem.
-    # Instead, we import functions from it
-    with add_to_sys_path(os.path.dirname(reduction_script)):
-        from reduce_REF_M import reduce_events_file, upload_html_report  # script being used as a module
+    #
+    # Invoke the main routine of the livereduction script. It will digest the autoreduction script
+    # reduce_REF_M.py we just created
+    #
+    accumulation_workspace = data_server.load_events("REF_M_42537.nxs.h5")
+    report_file = os.path.join(mock_filesystem.tempdir, "report.html")  # HTML report file
+    with mock.patch("mr_livereduce.reduce_REF_M_live_post_proc.GLOBAL_AR_DIR", mock_filesystem.tempdir):
+        with mock.patch("mr_livereduce.reduce_REF_M_live_post_proc.GLOBAL_LR_DIR", mock_filesystem.tempdir):
+            main(
+                accumulation_workspace,
+                outdir=mock_filesystem.tempdir,  # instead of /SNS/IPTS-31954/shared/autoreduce/
+                publish=False,  # don't upload the HTML report to the livedata server
+                report_file=report_file,
+            )
 
-        events_file = data_server.path_to("REF_M_42537.nxs.h5")
-        outdir = mock_filesystem.tempdir  # instead of /SNS/IPTS-31954/shared/autoreduce/
-        reports = reduce_events_file(events_file, outdir)  # reduce the two peaks and generate HTML reports
-        report_file = os.path.join(mock_filesystem.tempdir, "report.html")
-        upload_html_report(reports, publish=False, report_file=report_file)  # save reports to a files
+    #
+    # Check the results of the livereduction
+    #
 
-    # assert the HTML report has been created
-    assert os.path.isfile(report_file)
+    # assert the HTML report can be rendered by a headless Chromium browser
+    browser.get(f"file://{report_file}")
+    plotly_divs = browser.find_elements(By.CLASS_NAME, "plotly-graph-div")
+    assert len(plotly_divs) > 0, "No Plotly figures were rendered in the report."
 
     # assert reduction files have been produced for run 42537
-    for sn in (1, 2):  # peak number
+    for peak_number in (1, 2):
         for suffix in [
             "_Off_Off_autoreduce.dat",
             "_Off_Off_autoreduce.nxs.h5",
             "_On_Off_autoreduce.dat",
             "_On_Off_autoreduce.nxs.h5",
+            ".ort",  # ORSO ASCII format
             "_partial.py",
             ".json",
         ]:
-            file = f"REF_M_42537_{sn}{suffix}"
+            file = f"REF_M_42537_{peak_number}{suffix}"
             assert os.path.isfile(os.path.join(mock_filesystem.tempdir, file)), f"{file} doesn't exist"
 
     # assert stitched files have been produced (file names use run 42535 because this is
     # the first in the sequence of experiments encompassing run 42535 through 42538)
-    for sn in (1, 2):
-        for suffix in ["_combined.py", "_Off_Off_combined.dat", "_On_Off_combined.dat", "_tunable_combined.py"]:
-            file = f"REF_M_42535_{sn}{suffix}"
+    for peak_number in (1, 2):
+        for suffix in [
+            "_combined.py",
+            "_Off_Off_combined.dat",
+            "_On_Off_combined.dat",
+            "_combined.ort",  # ORSO ASCII format
+            "_tunable_combined.py",
+        ]:
+            file = f"REF_M_42535_{peak_number}{suffix}"
             assert os.path.isfile(os.path.join(mock_filesystem.tempdir, file)), f"{file} doesn't exist"